replaced Rewriter with TypeDeclVisitor
